front - move account outside of dashboard
change nav look
display company name even when there is only one company
add siret to company name in dashboard
refactor SideMenu and add a test
