+++ front/src/account/Account.scss
@@ -1,0 +1,41 @@
+.account {
+  display: flex;
+
+  .dashboard-content {
+    padding: 0 2em;
+  }
+
+  .account__form {
+    form {
+      max-width: 100%;
+      margin: 20px 0;
+    }
+
+    label {
+      margin-bottom: 0.5em;
+    }
+
+    .input__group {
+      align-items: stretch;
+
+      .icon-button {
+        padding: 0 0.7em;
+        margin: 0 0.7em;
+      }
+    }
+  }
+
+  .company__address {
+    font-style: normal;
+    margin-bottom: 1em;
+
+    p {
+      margin: 0;
+    }
+  }
+
+  .title {
+    margin-top: 1em;
+  }
+}
+
