+++ front/src/account/AccountMenu.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import SideMenu from "../common/SideMenu";
+import { NavLink, match } from "react-router-dom";
+
+export default function AccountMenu() {
+  return (
+    <SideMenu>
+      <NavLink to="/account">Paramètres du compte</NavLink>
+    </SideMenu>
+  );
+}
