+++ front/src/common/SideMenu.scss
@@ -1,0 +1,17 @@
+
+.dashboard-menu.side-menu {
+  background-color: #ebeff3;
+  flex: 0 1 auto;
+  padding: 0 20px 10px 0;
+  min-width: 240px;
+
+  .side-menu__mobile-trigger {
+    display: flex;
+    align-items: center;
+    font-size: 1.6em;
+    margin: 5px 10px;
+    cursor: pointer;
+  }
+}
+
+
